requirements.txt
aiogram>=3.15.0
aiohttp
aiohttp-socks
apscheduler
google-genai
python-dotenv
openmeteo-requests
requests-cache
retry-requests
sqlalchemy